Move Header test to App.test.tsx
